% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide_adjust-v2.0.R
\name{depth_adjust}
\alias{depth_adjust}
\title{Correction of depth measures for tide level}
\usage{
depth_adjust(
  tidetable_df,
  observed_df,
  time_tide_col = NULL,
  lev_tide_col = NULL,
  time_obs_col = NULL,
  depth_obs_col = NULL,
  depth = NULL,
  date.time = NULL,
  ts_tide_format = "\%d/\%m/\%Y \%H.\%M",
  ts_obs_format = ts_tide_format
)
}
\arguments{
\item{tidetable_df}{dataframe with tide table}

\item{observed_df}{dataframe with observed depth to be corrected.}

\item{time_tide_col}{either a number or a character specifying the position or the name of the column in `tidetable_df` for the time of the tide table}

\item{lev_tide_col}{either a number or a character specifying the position or the name of the column in `tidetable_df` for the tide level of the tide table}

\item{time_obs_col}{either a number or a character specifying the position or the name of the column in `observed_df` for the time of the observed depths}

\item{depth_obs_col}{either a number or a character specifying the position or the name of the column in `observed_df` for the observed depths}

\item{depth}{numeric. Single depth value to be adjusted}

\item{date.time}{Date and time of measurement of the depth than need adjustment.}

\item{ts_tide_format}{format used for time in `tidetable_df`}

\item{ts_obs_format}{format used for time in `observed_df`}
}
\description{
Correction of depth measures for tide level
}
